split entity users card
